Separated server/app to avoid test collisions (address in use)
