--- server/server.test.js
@@ -1,2 +1,0 @@
-const server = require('./server');
-
