+++ server/serverapp.test.js
@@ -1,0 +1,2 @@
+const server = require('./serverapp');
+
